pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Any4rd <Any4rd@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
djangorestframework = "^3.15.1"
drf-yasg = "^1.21.7"
psycopg2 = "^2.9.9"
environs = "^11.0.0"
django-environ = "^0.11.2"
django-cors-headers = "^4.3.1"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
